requirements.txt
# requirements.txt
streamlit==1.20.0
google-generativeai==0.3.2
python-docx==0.8.11
PyPDF2==3.0.1
Pillow==9.5.0
beautifulsoup4
requests
dotenv
piexif
scikit-learn==1.2.2